Add ability of any list item to open the edit popup
