+++ pomotrello/client/src/views/edit_modal_popup.js
@@ -1,0 +1,23 @@
+var EditModalPopup = function() {};
+
+EditModalPopup.prototype = {
+
+  editTaskModalPopUp: function() {
+    var editTaskModal = document.getElementById('edit-task-modal-popup');
+    var editTaskSpan = document.getElementById("close-edit-task-modal-popup");
+
+    // When the user clicks on <span> (x), close the modal
+    editTaskSpan.onclick = function() {
+      editTaskModal.style.display = "none";
+    }
+
+    // When the user clicks anywhere outside of the modal, close it
+    editTaskModal.onclick = function(event) {
+      if (event.target == editTaskModal) {
+        editTaskModal.style.display = "none";
+      }
+    }
+  }
+}
+
+module.exports = EditModalPopup;
